Move alternate implementations into benchmarks
